move AssetGraphComputation to its own module
branch-name: asset-graph-computation-separate-file
